+++ packages/component-library/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Editable Table
+- Horizontal Stacked Bar component.
 
 # v8.0.4 (April 21, 2026)